# openapi-schema.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/api/checked/abonners":{"post":{"operationId":"AppController_checkAbonner","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckAbonnerDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/substract/checked/byChecked/abonners":{"post":{"operationId":"AppController_debitAbonner","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PassageDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/signin":{"post":{"operationId":"AuthController_signin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/me":{"get":{"operationId":"AuthController_userInfo","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Auth"]}},"/api/user":{"post":{"operationId":"UserController_create","parameters":[{"name":"role","required":true,"in":"query","schema":{"enum":["USER","ADMIN","MANAGER","INFORMATICIEN"],"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Users"]},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["Users"]}},"/api/user/create/all":{"post":{"operationId":"UserController_createAll","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["Users"]}},"/api/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Users"]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/abonners":{"post":{"operationId":"AbonnersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AbonnerDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Abonners"]},"get":{"operationId":"AbonnersController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Abonner"}}}}}},"tags":["Abonners"]}},"/api/abonners/{id}":{"get":{"operationId":"AbonnersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Abonner"}}}}},"tags":["Abonners"]},"patch":{"operationId":"AbonnersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Abonner"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Abonners"]},"delete":{"operationId":"AbonnersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Abonners"]}},"/api/sites":{"post":{"operationId":"SitesController_create","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Site"}}}}},"tags":["Passages"]},"get":{"operationId":"SitesController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Site"}}}}}},"tags":["Passages"]}},"/api/sites/{site}/sum-par-voie":{"get":{"operationId":"SitesController_someMoneyByDateAndStreetOfSite","parameters":[{"name":"site","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VoieCount"}}}}}},"tags":["Passages"]}},"/api/sites/{voie}/sum":{"get":{"operationId":"SitesController_sumByStreet","parameters":[{"name":"voie","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VoieCount"}}}}},"tags":["Passages"]}},"/api/sites/sum/par-voie":{"get":{"operationId":"SitesController_sumStreet","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VoieCount"}}}}}},"tags":["Passages"]}},"/api/sites/sum/par-site":{"get":{"operationId":"SitesController_sumBySite","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SiteCount"}}}}}},"tags":["Passages"]}},"/api/sites/statistique/all":{"get":{"operationId":"SitesController_allStatist","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SiteCount"}}}}}},"tags":["Passages"]}},"/api/sites/{id}":{"get":{"operationId":"SitesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Site"}}}}},"tags":["Passages"]},"patch":{"operationId":"SitesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Passages"]},"delete":{"operationId":"SitesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Passages"]}},"/api/verifieCards":{"post":{"operationId":"VerifieCardsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckAbonnerDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifieCard"}}}}},"tags":["verifieCards"]},"get":{"operationId":"VerifieCardsController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VerifieCard"}}}}}},"tags":["verifieCards"]}},"/api/verifieCards/{id}":{"get":{"operationId":"VerifieCardsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifieCard"}}}}},"tags":["verifieCards"]},"patch":{"operationId":"VerifieCardsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["verifieCards"]}}},"info":{"title":"Validation API","description":"Système de gestion du pont de payage de èkpè :  Partie API","version":"1.0","contact":{}},"tags":[{"name":"cast","description":""}],"servers":[],"components":{"schemas":{"CheckAbonnerDto":{"type":"object","properties":{"idBadge":{"type":"string"}},"required":["idBadge"]},"PassageDto":{"type":"object","properties":{"site":{"type":"string"},"voie":{"type":"string"},"montant":{"type":"number"},"idBadge":{"type":"string"},"percepteur":{"type":"string"}},"required":["site","voie","montant","idBadge","percepteur"]},"LoginDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"UserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"enum":["USER","ADMIN","MANAGER","INFORMATICIEN"],"type":"string"},"account_number":{"type":"number"},"email_verified_at":{"format":"date-time","type":"string"}},"required":["name","email","password","role","account_number","email_verified_at"]},"User":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"enum":["USER","ADMIN","MANAGER","INFORMATICIEN"],"type":"string"},"api_token":{"type":"string"},"account_number":{"type":"number"},"remember_token":{"type":"string"},"email_verified_at":{"format":"date-time","type":"string"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"}},"required":["id","name","email","password","role","api_token","account_number","remember_token","email_verified_at","created_at","updated_at"]},"AbonnerDto":{"type":"object","properties":{"idBadge":{"type":"string"},"nom":{"type":"string"},"prenom":{"type":"string"},"tel":{"type":"string"},"adresse":{"type":"string"},"nip":{"type":"string"},"iduhf":{"type":"string"},"codeuhf":{"type":"string"},"plaque":{"type":"string"},"essieu":{"type":"string"},"solde":{"type":"number"},"type":{"type":"string"},"is_active":{"type":"boolean","default":true},"is_sentE":{"type":"boolean","default":false},"is_sentA":{"type":"boolean","default":false},"is_sentG":{"type":"boolean","default":false},"statut":{"type":"boolean"}},"required":["idBadge","nom","prenom","tel","adresse","nip","iduhf","codeuhf","plaque","essieu","solde","type","is_active","is_sentE","is_sentA","is_sentG","statut"]},"Abonner":{"type":"object","properties":{"id":{"type":"number"},"idBadge":{"type":"string"},"nom":{"type":"string"},"prenom":{"type":"string"},"tel":{"type":"string"},"adresse":{"type":"string"},"nip":{"type":"string"},"iduhf":{"type":"string"},"codeuhf":{"type":"string"},"plaque":{"type":"string"},"essieu":{"type":"string"},"solde":{"type":"number"},"type":{"type":"string"},"is_active":{"type":"boolean"},"is_sentE":{"type":"boolean"},"is_sentA":{"type":"boolean"},"is_sentG":{"type":"boolean"},"statut":{"type":"boolean"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"}},"required":["id","idBadge","nom","prenom","tel","adresse","nip","iduhf","codeuhf","plaque","essieu","solde","type","is_active","is_sentE","is_sentA","is_sentG","statut","created_at","updated_at"]},"Site":{"type":"object","properties":{"id":{"type":"number"},"site":{"type":"string"},"voie":{"type":"string"},"montant":{"type":"number"},"solde_initial":{"type":"number"},"montant_restant":{"type":"number"},"abonner_id":{"type":"number"},"idBadge":{"type":"string"},"is_sent":{"type":"boolean"},"is_sentE":{"type":"boolean"},"is_sentA":{"type":"boolean"},"is_sentG":{"type":"boolean"},"passage_verified_at":{"format":"date-time","type":"string"},"type_passage":{"type":"string"},"percepteur":{"type":"string"},"statut":{"type":"string"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"}},"required":["id","site","voie","montant","solde_initial","montant_restant","abonner_id","idBadge","is_sent","is_sentE","is_sentA","is_sentG","passage_verified_at","type_passage","percepteur","statut","created_at","updated_at"]},"VoieCount":{"type":"object","properties":{}},"SiteCount":{"type":"object","properties":{"site":{"type":"string"},"nbre":{"type":"number"},"sum":{"type":"number"},"voies":{"type":"array","items":{"$ref":"#/components/schemas/VoieCount"}}},"required":["nbre"]},"VerifieCard":{"type":"object","properties":{"id":{"type":"number"},"idBadge":{"type":"string"},"type_verifier":{"type":"string"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"}},"required":["id","idBadge","type_verifier","created_at","updated_at"]}}}}